+++ lib/src/core/systems/interval_entity_processing_system.dart
@@ -7,7 +7,7 @@ part of '../../../dartemis.dart';
 abstract class IntervalEntityProcessingSystem extends IntervalEntitySystem {
   /// Create a new [IntervalEntityProcessingSystem]. It requires at least one
   /// component.
-  IntervalEntityProcessingSystem(int super.interval, super.aspect);
+  IntervalEntityProcessingSystem(super.interval, super.aspect);
 
   /// Process an [entity] this system is interested in.
   void processEntity(int entity);