+++ lib/src/core/systems/interval_entity_system.dart
@@ -3,20 +3,21 @@ part of '../../../dartemis.dart';
 /// A system that processes entities at a interval in milliseconds.
 /// A typical usage would be a collision system or physics system.
 abstract class IntervalEntitySystem extends EntitySystem {
-  num _acc = 0;
-  num _delta = 0;
-  final num _interval;
+  double _acc = 0;
+  double _intervalDelta = 0;
+  final double _interval;
 
   /// Create an [IntervalEntitySystem] with the specified interval and [aspect].
   IntervalEntitySystem(this._interval, Aspect aspect) : super(aspect);
 
   /// Returns the accumulated delta since the system was last invoked.
-  num get delta => _delta;
+  @override
+  double get delta => _intervalDelta;
 
   @override
   bool checkProcessing() {
     _acc += world.delta;
-    _delta += world.delta;
+    _intervalDelta += world.delta;
     if (_acc >= _interval) {
       _acc -= _interval;
       return true;
@@ -29,6 +30,6 @@ abstract class IntervalEntitySystem extends EntitySystem {
   /// Call `super.end()` if you overwrite this function.
   @override
   void end() {
-    _delta = 0;
+    _intervalDelta = 0;
   }
 }